UI-16:  Made changes to project as a result of implementing a solution in the Message_DTO_Model project to capture the resource type name property in the response message pertaining to the task of getting a list of resource sub types.

+++ amh-parent/amh_security/src/test/java/org/rmt2/api/handler/admin/UserQueryMessageHandlerTest.java
@@ -115,7 +115,7 @@ public class UserQueryMessageHandlerTest extends BaseAuthenticationMessageHandle
         
         for (int ndx = 0; ndx < actualRepsonse.getProfile().getUserInfo().size(); ndx++) {
             UserType a = actualRepsonse.getProfile().getUserInfo().get(ndx);
-            Assert.assertEquals(SecurityMockOrmDataFactory.TEST_USER_ID + ndx, a.getLoginId().intValue(), 0);
+            Assert.assertEquals(SecurityMockOrmDataFactory.TEST_USER_ID + ndx, a.getLoginId(), 0);
         }
     }
